changes for lineitem apis

+++ src/main/java/com/facilio/bmsconsole/commands/AddOrUpdatePurchaseOrderCommand.java
@@ -8,6 +8,7 @@ import com.facilio.bmsconsole.context.PurchaseOrderLineItemContext;
 import com.facilio.bmsconsole.context.ReceivableContext;
 import com.facilio.bmsconsole.context.ReceivableContext.Status;
 import com.facilio.bmsconsole.util.LocationAPI;
+import com.facilio.bmsconsole.util.RecordAPI;
 import com.facilio.constants.FacilioConstants;
 import com.facilio.db.criteria.CriteriaAPI;
 import com.facilio.db.criteria.operators.NumberOperators;
@@ -36,7 +37,7 @@ public class AddOrUpdatePurchaseOrderCommand implements Command {
 			
 			FacilioModule lineModule = modBean.getModule(FacilioConstants.ContextNames.PURCHASE_ORDER_LINE_ITEMS);
 			
-			if (CollectionUtils.isEmpty(purchaseOrderContext.getLineItems())) {
+			if (purchaseOrderContext.getId() <= 0 && CollectionUtils.isEmpty(purchaseOrderContext.getLineItems())) {
 				throw new Exception("Line items cannot be empty");
 			}
 			
@@ -53,28 +54,33 @@ public class AddOrUpdatePurchaseOrderCommand implements Command {
 			purchaseOrderContext.setShipToAddress(LocationAPI.getPoPrLocation(purchaseOrderContext.getStoreRoom(), purchaseOrderContext.getShipToAddress(), "SHIP_TO_Location", true));
 			purchaseOrderContext.setBillToAddress(LocationAPI.getPoPrLocation(purchaseOrderContext.getVendor(), purchaseOrderContext.getBillToAddress(), "BILL_TO_Location", false));
 			if (purchaseOrderContext.getId() > 0) {
-				updateRecord(purchaseOrderContext, module, fields);
-				
-				DeleteRecordBuilder<PurchaseOrderLineItemContext> deleteBuilder = new DeleteRecordBuilder<PurchaseOrderLineItemContext>()
-						.module(lineModule)
-						.andCondition(CriteriaAPI.getCondition("PO_ID", "poid", String.valueOf(purchaseOrderContext.getId()), NumberOperators.EQUALS));
-				deleteBuilder.delete();
+				RecordAPI.updateRecord(purchaseOrderContext, module, fields);
+				if(purchaseOrderContext.getLineItems() != null) {
+					DeleteRecordBuilder<PurchaseOrderLineItemContext> deleteBuilder = new DeleteRecordBuilder<PurchaseOrderLineItemContext>()
+							.module(lineModule)
+							.andCondition(CriteriaAPI.getCondition("PO_ID", "poid", String.valueOf(purchaseOrderContext.getId()), NumberOperators.EQUALS));
+					deleteBuilder.delete();
+					updateLineItems(purchaseOrderContext);
+					RecordAPI.addRecord(false,purchaseOrderContext.getLineItems(), lineModule, modBean.getAllFields(lineModule.getName()));
+			
+				}
 			} else {
 				
 				if(purchaseOrderContext.getOrderedTime() == -1) {
 					purchaseOrderContext.setOrderedTime(System.currentTimeMillis());
 				}
 				purchaseOrderContext.setStatus(PurchaseOrderContext.Status.REQUESTED);
-				addRecord(true,Collections.singletonList(purchaseOrderContext), module, fields);
+				RecordAPI.addRecord(true,Collections.singletonList(purchaseOrderContext), module, fields);
 				FacilioModule receivableModule = modBean.getModule(FacilioConstants.ContextNames.RECEIVABLE);
 				ReceivableContext receivableContext = new ReceivableContext();
 				receivableContext.setPoId(purchaseOrderContext);
 				receivableContext.setStatus(Status.YET_TO_RECEIVE);
-				addRecord(true,Collections.singletonList(receivableContext), receivableModule, modBean.getAllFields(receivableModule.getName()));
+				RecordAPI.addRecord(true,Collections.singletonList(receivableContext), receivableModule, modBean.getAllFields(receivableModule.getName()));
+				updateLineItems(purchaseOrderContext);
+				RecordAPI.addRecord(false,purchaseOrderContext.getLineItems(), lineModule, modBean.getAllFields(lineModule.getName()));
+		
 			}
 			
-			updateLineItems(purchaseOrderContext);
-			addRecord(false,purchaseOrderContext.getLineItems(), lineModule, modBean.getAllFields(lineModule.getName()));
 			
 			context.put(FacilioConstants.ContextNames.RECORD, purchaseOrderContext);
 		}
@@ -88,24 +94,7 @@ public class AddOrUpdatePurchaseOrderCommand implements Command {
 		}
 	}
 	
-	private void addRecord(boolean isLocalIdNeeded, List<? extends ModuleBaseWithCustomFields> list, FacilioModule module, List<FacilioField> fields) throws Exception {
-		InsertRecordBuilder insertRecordBuilder = new InsertRecordBuilder<>()
-				.module(module)
-				.fields(fields);
-		if(isLocalIdNeeded) {
-			insertRecordBuilder.withLocalId();
-		}
-		insertRecordBuilder.addRecords(list);
-		insertRecordBuilder.save();
-	}
-	
-	public void updateRecord(ModuleBaseWithCustomFields data, FacilioModule module, List<FacilioField> fields) throws Exception {
-		UpdateRecordBuilder updateRecordBuilder = new UpdateRecordBuilder<ModuleBaseWithCustomFields>()
-				.module(module)
-				.fields(fields)
-				.andCondition(CriteriaAPI.getIdCondition(data.getId(), module));
-		updateRecordBuilder.update(data);
-	}
+
 	private void updateLineItemCost(PurchaseOrderLineItemContext lineItemContext){
 		if(lineItemContext.getUnitPrice() > 0) {
 		  lineItemContext.setCost(lineItemContext.getUnitPrice() * lineItemContext.getQuantity());
